extend dataset registration UI for S3
