+++ frontend/package.json
@@ -14,6 +14,8 @@
   "dependencies": {
     "@cznethub/cznet-vue-core": "^0.2.3",
     "@unhead/vue": "^1.8.14",
+    "@vuelidate/core": "^2.0.3",
+    "@vuelidate/validators": "^2.0.4",
     "@vueuse/core": "^10.9.0",
     "@vueuse/head": "^2.0.0",
     "@vuex-orm/core": "^0.36.4",
